Bug fixing
A Bug's Life is a 1998 American computer-animated comedy film produced by Pixar Animation Studios for Walt Disney Pictures. It was the second film produced by Pixar. Directed by John Lasseter and co-directed and written by Andrew Stanton, the film involves a misfit ant, Flik, who is looking for "tough warriors" to save his colony from a protection racket run by Hopper's gang of grasshoppers. Unfortunately, the "warriors" he brings back turn out to be an inept troupe of Circus Bugs.

The film was initially inspired by Aesop's fable The Ant and the Grasshopper.[3][4] Production began shortly after the release of Toy Story in 1995. The screenplay was penned by Stanton and comedy writers Donald McEnery and Bob Shaw from a story by Lasseter, Stanton, and Ranft. The ants in the film were redesigned to be more appealing, and Pixar's animation unit employed technical innovations in computer animation. Randy Newman composed the music for the film. During production, a controversial public feud erupted between Steve Jobs and Lasseter of Pixar and DreamWorks co-founder Jeffrey Katzenberg due to the parallel production of his similar film Antz, which was released the same year.

The film was released on November 20, 1998, received positive reviews and grossed $363 million at the box office. It was the first film to be digitally transferred frame-by-frame and released on DVD, and has been released multiple times on home video.